pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-llm-eval"
version = "0.1.0"
description = "Mini LLM evaluation runner"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.7,<3",
    "PyYAML>=6.0,<7",
    "httpx>=0.25,<1",
    "aiosqlite>=0.19,<1",
    "typer>=0.12,<1",
    "rich>=13,<14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5,<6",
]

[project.scripts]
mini-llm-eval = "mini_llm_eval.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
